+++ db_conn/tmp_tables.py
@@ -45,7 +45,7 @@ def merge_extract():
 
         cur.execute("""CREATE TEMP TABLE temp_merge (filepath varchar, workspace varchar, dataset varchar,collection varchar,filename varchar,imagerytype varchar,imagetype varchar,
                                                      startdepth numeric, enddepth numeric, startbox smallint, endbox smallint, status varchar, flag_error varchar,condition_error varchar,
-                                                     exif smallint, width smallint, height smallint ,img_obs varchar,project_code varchar );""")
+                                                     exif smallint, width bigint, height bigint ,img_obs varchar,project_code varchar );""")
         
         print('temp_merge created...')
